requirements.txt
kivy[base]==2.2.0
kivymd==1.1.1
pycryptodome==3.19.0
plyer==2.1.0
python-for-android
buildozer
